+++ sample/src/main/java/com/github/panpf/sketch/sample/ui/viewer/ImageFragment.kt
@@ -19,6 +19,7 @@ import android.os.Bundle
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.navArgs
 import com.github.panpf.assemblyadapter.pager.FragmentItemFactory
+import com.github.panpf.sketch.cache.CachePolicy.DISABLED
 import com.github.panpf.sketch.displayImage
 import com.github.panpf.sketch.sample.databinding.ImageFragmentBinding
 import com.github.panpf.sketch.sample.model.ImageDetail
@@ -34,6 +35,9 @@ class ImageFragment : BindingFragment<ImageFragmentBinding>() {
         binding.imageImage.apply {
             showDataFromLogo()
             displayImage(args.url) {
+                memoryCachePolicy(DISABLED)
+                resultCachePolicy(DISABLED)
+                downloadCachePolicy(DISABLED)
                 lifecycle(viewLifecycleOwner.lifecycle)
                 listener(
                     onStart = {